% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{e_size}
\alias{e_size}
\alias{e_size,DNAStringSet-method}
\alias{e_size,character-method}
\title{Generate \code{E-size} metric.
For details see methods for https://dx.doi.org/10.1101\%2Fgr.131383.111}
\usage{
e_size(obj)

\S4method{e_size}{DNAStringSet}(obj)

\S4method{e_size}{character}(obj)
}
\arguments{
\item{obj}{character, DNAStringSet}
}
\description{
Generate \code{E-size} metric.
For details see methods for https://dx.doi.org/10.1101\%2Fgr.131383.111
}
