% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{NX0}
\alias{NX0}
\alias{NX0,DNAStringSet-method}
\alias{NX0,character-method}
\title{Generate N50 or related metrics.
generate N50, N90, etc.}
\usage{
NX0(obj, percentile = 50L)

\S4method{NX0}{DNAStringSet}(obj, percentile = 50L)

\S4method{NX0}{character}(obj, percentile = 50L)
}
\arguments{
\item{obj}{character, DNAStringSet}
}
\description{
Generate N50 or related metrics.
generate N50, N90, etc.
}
