% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contiguity.R
\name{nx_curve}
\alias{nx_curve}
\alias{multi_nx_curve}
\title{Generate auN curve
Inspired by http://lh3.github.io/2020/04/08/a-new-metric-on-assembly-contiguity}
\usage{
nx_curve(obj, percentiles = as.integer(seq(0L, 100L, 10L)))

multi_nx_curve(objs, percentiles = seq(0L, 100L, 10L))
}
\arguments{
\item{obj}{DNAStringSet, character}
}
\description{
Generates N10, N20 ... N100 rather than NG*.
}
